New font bitmap decompression code and other tweaks to the debug text rendering.
Thanks to the changes made to font-tool, we now have a much better
compression format for the font glyph bitmap. It is now stored with a
custom LZW-based compression, so the embedded data array is
significantly smaller. It is now also stored as an escaped hexadecimal
string, so the code representation itself is smaller, which has shaved
about 600+ lines from the implementation source!

The new font-tool now does a much better job exporting the grayscale
bitmap, so the quality of the debug text rendering is also greatly
improved in account of that.

The only negative is that the decompression code is now larger and more
complex, but it is still very efficient and doesn’t allocate any more
memory than what the previous RLE decoder already did.

I also tightened the char drawing functions to remove redundant vars
and hoist a couple loop invariants. This should help lower-end
compilers generate better code.